refactor: move compaction to context.py, remove .bak
CompactionController lives in context.py alongside ContextManager —
all context concerns in one file. Deleted standalone compaction.py
and app.py.bak safety net.
